requirements.txt
flask
imutils
opencv-python-headless
Pillow
scikit-image
gunicorn